reorganize test dir
