+++ tests/integration/cwd-option.spec.ts
