+++ tests/unit/depngn/getDependencies/mocks/withPackageJson/package.json
