+++ tests/unit/depngn/getDependencies/mocks/withoutPackageJson/notPackage.json
